anton-rx-backend/requirements.txt
google-genai>=1.0.0
pdfplumber>=0.11.0
rich>=13.7.0
